mqtt, analog, fahrenheit and other changes

+++ src/web/WebCustomizationService.h
@@ -44,7 +44,7 @@ class AnalogCustomization {
   public:
     uint8_t     id;
     std::string name;
-    uint16_t    offset;
+    float       offset;
     float       factor;
     uint8_t     uom;  // 0 is none
     int8_t      type; // -1 is for deletion
